refactor of integration tests
