+++ crates/aos-testkit/tests/world_integration.rs
@@ -1,6 +1,6 @@
 use aos_air_exec::Value as ExprValue;
 use aos_air_types::{
-    DefPlan, EffectKind, Expr, ExprConst, ExprRecord, HashRef, PlanBind, PlanBindEffect, PlanEdge,
+    DefPlan, EffectKind, Expr, ExprConst, ExprRecord, PlanBind, PlanBindEffect, PlanEdge,
     PlanStep, PlanStepAssign, PlanStepAwaitEvent, PlanStepAwaitReceipt, PlanStepEmitEffect,
     PlanStepEnd, PlanStepKind, PlanStepRaiseEvent,
 };
@@ -9,221 +9,14 @@ use aos_effects::builtins::{
 };
 use aos_effects::{EffectReceipt, ReceiptStatus};
 use aos_kernel::error::KernelError;
-use aos_kernel::journal::mem::MemJournal;
 use aos_testkit::fixtures::{self, START_SCHEMA};
-use aos_testkit::{TestStore, TestWorld, effect_params_text};
+use aos_testkit::{TestWorld, effect_params_text, fake_hash};
 use aos_wasm_abi::{ReducerEffect, ReducerOutput};
 use indexmap::IndexMap;
 use serde_cbor;
-use std::sync::Arc;
 
-fn fulfillment_manifest(store: &Arc<TestStore>) -> aos_kernel::manifest::LoadedManifest {
-    let result_module = fixtures::stub_reducer_module(
-        store,
-        "com.acme/ResultReducer@1",
-        &ReducerOutput {
-            state: Some(vec![0xEE]),
-            domain_events: vec![],
-            effects: vec![],
-            ann: None,
-        },
-    );
-
-    let plan_name = "com.acme/Fulfill@1".to_string();
-    let plan = DefPlan {
-        name: plan_name.clone(),
-        input: fixtures::schema("com.acme/PlanIn@1"),
-        output: None,
-        locals: IndexMap::new(),
-        steps: vec![
-            PlanStep {
-                id: "emit".into(),
-                kind: PlanStepKind::EmitEffect(PlanStepEmitEffect {
-                    kind: EffectKind::HttpRequest,
-                    params: fixtures::text_expr("body"),
-                    cap: "cap_http".into(),
-                    bind: PlanBindEffect {
-                        effect_id_as: "req".into(),
-                    },
-                }),
-            },
-            PlanStep {
-                id: "await".into(),
-                kind: PlanStepKind::AwaitReceipt(PlanStepAwaitReceipt {
-                    for_expr: fixtures::var_expr("req"),
-                    bind: PlanBind { var: "resp".into() },
-                }),
-            },
-            PlanStep {
-                id: "raise".into(),
-                kind: PlanStepKind::RaiseEvent(PlanStepRaiseEvent {
-                    reducer: result_module.name.clone(),
-                    event: Expr::Record(ExprRecord {
-                        record: IndexMap::from([
-                            ("$schema".into(), fixtures::text_expr("com.acme/Result@1")),
-                            ("value".into(), Expr::Const(ExprConst::Int { int: 9 })),
-                        ]),
-                    }),
-                    key: None,
-                }),
-            },
-            PlanStep {
-                id: "end".into(),
-                kind: PlanStepKind::End(PlanStepEnd { result: None }),
-            },
-        ],
-        edges: vec![
-            PlanEdge {
-                from: "emit".into(),
-                to: "await".into(),
-                when: None,
-            },
-            PlanEdge {
-                from: "await".into(),
-                to: "raise".into(),
-                when: None,
-            },
-            PlanEdge {
-                from: "raise".into(),
-                to: "end".into(),
-                when: None,
-            },
-        ],
-        required_caps: vec!["cap_http".into()],
-        allowed_effects: vec![EffectKind::HttpRequest],
-        invariants: vec![],
-    };
-
-    let routing = vec![fixtures::routing_event(
-        "com.acme/Result@1",
-        &result_module.name,
-    )];
-    fixtures::build_loaded_manifest(
-        vec![plan],
-        vec![fixtures::start_trigger(&plan_name)],
-        vec![result_module],
-        routing,
-    )
-}
-
-fn await_event_manifest(store: &Arc<TestStore>) -> aos_kernel::manifest::LoadedManifest {
-    let result_module = fixtures::stub_reducer_module(
-        store,
-        "com.acme/EventResult@1",
-        &ReducerOutput {
-            state: Some(vec![0xAB]),
-            domain_events: vec![],
-            effects: vec![],
-            ann: None,
-        },
-    );
-    let unblock_event = fixtures::domain_event(
-        "com.acme/Unblock@1",
-        &fixtures::plan_input_record(vec![]),
-    );
-    let unblock_emitter = fixtures::stub_event_emitting_reducer(
-        store,
-        "com.acme/UnblockEmitter@1",
-        vec![unblock_event],
-    );
-
-    let plan_name = "com.acme/WaitForEvent@1".to_string();
-    let plan = DefPlan {
-        name: plan_name.clone(),
-        input: fixtures::schema("com.acme/PlanIn@1"),
-        output: None,
-        locals: IndexMap::new(),
-        steps: vec![
-            PlanStep {
-                id: "await".into(),
-                kind: PlanStepKind::AwaitEvent(PlanStepAwaitEvent {
-                    event: fixtures::schema("com.acme/Unblock@1"),
-                    where_clause: None,
-                    bind: PlanBind { var: "evt".into() },
-                }),
-            },
-            PlanStep {
-                id: "raise".into(),
-                kind: PlanStepKind::RaiseEvent(PlanStepRaiseEvent {
-                    reducer: result_module.name.clone(),
-                    key: None,
-                    event: Expr::Record(ExprRecord {
-                        record: IndexMap::from([
-                            (
-                                "$schema".into(),
-                                fixtures::text_expr("com.acme/EventDone@1"),
-                            ),
-                            ("value".into(), Expr::Const(ExprConst::Int { int: 5 })),
-                        ]),
-                    }),
-                }),
-            },
-            PlanStep {
-                id: "end".into(),
-                kind: PlanStepKind::End(PlanStepEnd { result: None }),
-            },
-        ],
-        edges: vec![
-            PlanEdge {
-                from: "await".into(),
-                to: "raise".into(),
-                when: None,
-            },
-            PlanEdge {
-                from: "raise".into(),
-                to: "end".into(),
-                when: None,
-            },
-        ],
-        required_caps: vec![],
-        allowed_effects: vec![],
-        invariants: vec![],
-    };
-
-    let routing = vec![
-        fixtures::routing_event("com.acme/EventDone@1", &result_module.name),
-        fixtures::routing_event("com.acme/EmitUnblock@1", &unblock_emitter.name),
-    ];
-    fixtures::build_loaded_manifest(
-        vec![plan],
-        vec![fixtures::start_trigger(&plan_name)],
-        vec![result_module, unblock_emitter],
-        routing,
-    )
-}
-
-fn timer_manifest(store: &Arc<TestStore>) -> aos_kernel::manifest::LoadedManifest {
-    let timer_output = ReducerOutput {
-        state: Some(vec![0x01]),
-        domain_events: vec![],
-        effects: vec![ReducerEffect::new(
-            "timer.set",
-            serde_cbor::to_vec(&TimerSetParams {
-                deliver_at_ns: 5,
-                key: Some("retry".into()),
-            })
-            .unwrap(),
-        )],
-        ann: None,
-    };
-    let timer_emitter =
-        fixtures::stub_reducer_module(store, "com.acme/TimerEmitter@1", &timer_output);
-
-    let handler_output = ReducerOutput {
-        state: Some(vec![0xCC]),
-        domain_events: vec![],
-        effects: vec![],
-        ann: None,
-    };
-    let timer_handler =
-        fixtures::stub_reducer_module(store, "com.acme/TimerHandler@1", &handler_output);
-
-    let routing = vec![
-        fixtures::routing_event(fixtures::SYS_TIMER_FIRED, &timer_handler.name),
-        fixtures::routing_event(START_SCHEMA, &timer_emitter.name),
-    ];
-    fixtures::build_loaded_manifest(vec![], vec![], vec![timer_emitter, timer_handler], routing)
-}
+mod helpers;
+use helpers::timer_manifest;
 
 /// Happy-path end-to-end: reducer emits an intent, plan does work, receipt feeds a result event
 /// back into the reducer. Mirrors the “single plan orchestration” pattern in the spec.
@@ -342,52 +135,6 @@ fn single_plan_orchestration_completes_after_receipt() {
     assert_eq!(
         world.kernel.reducer_state("com.acme/ResultReducer@1"),
         Some(&vec![0xEE])
-    );
-}
-
-#[test]
-fn journal_replay_restores_state() {
-    let store = fixtures::new_mem_store();
-    let manifest_run = fulfillment_manifest(&store);
-    let mut world = TestWorld::with_store(store.clone(), manifest_run).unwrap();
-
-    let input = fixtures::plan_input_record(vec![("id", ExprValue::Text("123".into()))]);
-    world.submit_event_value(START_SCHEMA, &input);
-    world.tick_n(2).unwrap();
-
-    let mut effects = world.drain_effects();
-    assert_eq!(effects.len(), 1);
-    let effect = effects.remove(0);
-    let receipt_payload = serde_cbor::to_vec(&ExprValue::Text("done".into())).unwrap();
-    let receipt = EffectReceipt {
-        intent_hash: effect.intent_hash,
-        adapter_id: "adapter.http".into(),
-        status: ReceiptStatus::Ok,
-        payload_cbor: receipt_payload,
-        cost_cents: None,
-        signature: vec![],
-    };
-    world.kernel.handle_receipt(receipt).unwrap();
-    world.kernel.tick_until_idle().unwrap();
-
-    let final_state = world
-        .kernel
-        .reducer_state("com.acme/ResultReducer@1")
-        .cloned()
-        .unwrap();
-    let journal_entries = world.kernel.dump_journal().unwrap();
-
-    let manifest_replay = fulfillment_manifest(&store);
-    let replay_journal = MemJournal::from_entries(&journal_entries);
-    let replay_world =
-        TestWorld::with_store_and_journal(store.clone(), manifest_replay, Box::new(replay_journal))
-            .unwrap();
-
-    assert_eq!(
-        replay_world
-            .kernel
-            .reducer_state("com.acme/ResultReducer@1"),
-        Some(&final_state)
     );
 }
 
@@ -533,260 +280,6 @@ fn reducer_timer_receipt_routes_event_to_handler() {
     };
     let err = world.kernel.handle_receipt(unknown).unwrap_err();
     assert!(matches!(err, KernelError::UnknownReceipt(_)));
-}
-
-#[test]
-fn reducer_timer_receipt_replays_from_journal() {
-    let store = fixtures::new_mem_store();
-    let manifest = timer_manifest(&store);
-    let mut world = TestWorld::with_store(store.clone(), manifest).unwrap();
-    world.submit_event_value(START_SCHEMA, &fixtures::plan_input_record(vec![]));
-    world.tick_n(1).unwrap();
-
-    let effect = world.drain_effects().pop().expect("timer effect");
-    let receipt = EffectReceipt {
-        intent_hash: effect.intent_hash,
-        adapter_id: "adapter.timer".into(),
-        status: ReceiptStatus::Ok,
-        payload_cbor: serde_cbor::to_vec(&TimerSetReceipt {
-            delivered_at_ns: 10,
-            key: Some("retry".into()),
-        })
-        .unwrap(),
-        cost_cents: Some(1),
-        signature: vec![1, 2, 3],
-    };
-    world.kernel.handle_receipt(receipt).unwrap();
-    world.tick_n(1).unwrap();
-
-    let final_state = world
-        .kernel
-        .reducer_state("com.acme/TimerHandler@1")
-        .cloned()
-        .unwrap();
-    let journal_entries = world.kernel.dump_journal().unwrap();
-
-    let replay_world = TestWorld::with_store_and_journal(
-        store.clone(),
-        timer_manifest(&store),
-        Box::new(MemJournal::from_entries(&journal_entries)),
-    )
-    .unwrap();
-
-    assert_eq!(
-        replay_world.kernel.reducer_state("com.acme/TimerHandler@1"),
-        Some(&final_state)
-    );
-}
-
-#[test]
-fn plan_snapshot_resumes_after_receipt() {
-    let store = fixtures::new_mem_store();
-    let manifest = fulfillment_manifest(&store);
-    let mut world = TestWorld::with_store(store.clone(), manifest).unwrap();
-
-    let input = fixtures::plan_input_record(vec![("id", ExprValue::Text("123".into()))]);
-    world.submit_event_value(START_SCHEMA, &input);
-    world.tick_n(2).unwrap();
-
-    let effect = world
-        .drain_effects()
-        .pop()
-        .expect("expected effect before snapshot");
-
-    world.kernel.create_snapshot().unwrap();
-    let journal_entries = world.kernel.dump_journal().unwrap();
-
-    let mut replay_world = TestWorld::with_store_and_journal(
-        store.clone(),
-        fulfillment_manifest(&store),
-        Box::new(MemJournal::from_entries(&journal_entries)),
-    )
-    .unwrap();
-
-    let receipt_payload = serde_cbor::to_vec(&ExprValue::Text("done".into())).unwrap();
-    let receipt = EffectReceipt {
-        intent_hash: effect.intent_hash,
-        adapter_id: "adapter.http".into(),
-        status: ReceiptStatus::Ok,
-        payload_cbor: receipt_payload,
-        cost_cents: None,
-        signature: vec![],
-    };
-    replay_world.kernel.handle_receipt(receipt).unwrap();
-    replay_world.kernel.tick_until_idle().unwrap();
-
-    assert_eq!(
-        replay_world
-            .kernel
-            .reducer_state("com.acme/ResultReducer@1"),
-        Some(&vec![0xEE])
-    );
-}
-
-#[test]
-fn plan_snapshot_preserves_effect_queue() {
-    let store = fixtures::new_mem_store();
-    let manifest = fulfillment_manifest(&store);
-    let mut world = TestWorld::with_store(store.clone(), manifest).unwrap();
-
-    let input = fixtures::plan_input_record(vec![("id", ExprValue::Text("123".into()))]);
-    world.submit_event_value(START_SCHEMA, &input);
-    world.tick_n(2).unwrap();
-
-    world.kernel.create_snapshot().unwrap();
-    let entries = world.kernel.dump_journal().unwrap();
-
-    let mut replay_world = TestWorld::with_store_and_journal(
-        store.clone(),
-        fulfillment_manifest(&store),
-        Box::new(MemJournal::from_entries(&entries)),
-    )
-    .unwrap();
-
-    let mut intents = replay_world.drain_effects();
-    assert_eq!(intents.len(), 1, "effect queue should persist across snapshot");
-    let effect = intents.remove(0);
-
-    let receipt_payload = serde_cbor::to_vec(&ExprValue::Text("done".into())).unwrap();
-    let receipt = EffectReceipt {
-        intent_hash: effect.intent_hash,
-        adapter_id: "adapter.http".into(),
-        status: ReceiptStatus::Ok,
-        payload_cbor: receipt_payload,
-        cost_cents: None,
-        signature: vec![],
-    };
-    replay_world.kernel.handle_receipt(receipt).unwrap();
-    replay_world.kernel.tick_until_idle().unwrap();
-
-    assert_eq!(
-        replay_world
-            .kernel
-            .reducer_state("com.acme/ResultReducer@1"),
-        Some(&vec![0xEE])
-    );
-}
-
-#[test]
-fn plan_snapshot_resumes_after_event() {
-    let store = fixtures::new_mem_store();
-    let manifest = await_event_manifest(&store);
-    let mut world = TestWorld::with_store(store.clone(), manifest).unwrap();
-
-    let input = fixtures::plan_input_record(vec![("id", ExprValue::Text("evt".into()))]);
-    world.submit_event_value(START_SCHEMA, &input);
-    world.tick_n(1).unwrap();
-    world.tick_n(1).unwrap();
-
-    world.kernel.create_snapshot().unwrap();
-    let entries = world.kernel.dump_journal().unwrap();
-
-    let mut replay_world = TestWorld::with_store_and_journal(
-        store.clone(),
-        await_event_manifest(&store),
-        Box::new(MemJournal::from_entries(&entries)),
-    )
-    .unwrap();
-
-    let unblock = fixtures::plan_input_record(vec![]);
-    replay_world.submit_event_value("com.acme/EmitUnblock@1", &unblock);
-    replay_world.kernel.tick_until_idle().unwrap();
-
-    assert_eq!(
-        replay_world.kernel.reducer_state("com.acme/EventResult@1"),
-        Some(&vec![0xAB])
-    );
-}
-
-#[test]
-fn reducer_timer_snapshot_resumes_on_receipt() {
-    let store = fixtures::new_mem_store();
-    let manifest = timer_manifest(&store);
-    let mut world = TestWorld::with_store(store.clone(), manifest).unwrap();
-
-    world.submit_event_value(START_SCHEMA, &fixtures::plan_input_record(vec![]));
-    world.tick_n(1).unwrap();
-
-    let effect = world
-        .drain_effects()
-        .pop()
-        .expect("expected timer effect before snapshot");
-
-    world.kernel.create_snapshot().unwrap();
-    let entries = world.kernel.dump_journal().unwrap();
-
-    let mut replay_world = TestWorld::with_store_and_journal(
-        store.clone(),
-        timer_manifest(&store),
-        Box::new(MemJournal::from_entries(&entries)),
-    )
-    .unwrap();
-
-    let receipt = EffectReceipt {
-        intent_hash: effect.intent_hash,
-        adapter_id: "adapter.timer".into(),
-        status: ReceiptStatus::Ok,
-        payload_cbor: serde_cbor::to_vec(&TimerSetReceipt {
-            delivered_at_ns: 10,
-            key: Some("retry".into()),
-        })
-        .unwrap(),
-        cost_cents: Some(1),
-        signature: vec![1, 2, 3],
-    };
-    replay_world.kernel.handle_receipt(receipt).unwrap();
-    replay_world.tick_n(2).unwrap();
-
-    assert_eq!(
-        replay_world.kernel.reducer_state("com.acme/TimerHandler@1"),
-        Some(&vec![0xCC])
-    );
-}
-
-#[test]
-fn snapshot_replay_restores_state() {
-    fn build_manifest(store: &Arc<TestStore>) -> aos_kernel::manifest::LoadedManifest {
-        let reducer = fixtures::stub_reducer_module(
-            store,
-            "com.acme/Simple@1",
-            &ReducerOutput {
-                state: Some(vec![0xAA]),
-                domain_events: vec![],
-                effects: vec![],
-                ann: None,
-            },
-        );
-        let routing = vec![fixtures::routing_event(START_SCHEMA, &reducer.name)];
-        fixtures::build_loaded_manifest(vec![], vec![], vec![reducer], routing)
-    }
-
-    let store = fixtures::new_mem_store();
-    let manifest = build_manifest(&store);
-    let mut world = TestWorld::with_store(store.clone(), manifest).unwrap();
-    world.submit_event_value(START_SCHEMA, &fixtures::plan_input_record(vec![]));
-    world.tick_n(1).unwrap();
-
-    world.kernel.create_snapshot().unwrap();
-
-    let final_state = world
-        .kernel
-        .reducer_state("com.acme/Simple@1")
-        .cloned()
-        .unwrap();
-    let entries = world.kernel.dump_journal().unwrap();
-
-    let replay_world = TestWorld::with_store_and_journal(
-        store.clone(),
-        build_manifest(&store),
-        Box::new(MemJournal::from_entries(&entries)),
-    )
-    .unwrap();
-
-    assert_eq!(
-        replay_world.kernel.reducer_state("com.acme/Simple@1"),
-        Some(&final_state)
-    );
 }
 
 /// Guards on plan edges should gate side-effects and completion state.
@@ -1393,9 +886,4 @@ fn raised_events_are_routed_to_reducers() {
         world.kernel.reducer_state("com.acme/Reducer@1"),
         Some(&vec![0xEE])
     );
-}
-
-fn fake_hash(byte: u8) -> HashRef {
-    let hex = format!("{:02x}", byte);
-    HashRef::new(format!("sha256:{}", hex.repeat(32))).unwrap()
 }
